+++ NEWS.md
@@ -1,3 +1,7 @@
+## Version `v2.0.14` (07.08.2025)
+### Changes:
+* Remove HTTP priority header to work around DoS CVE-2025-31650 ([MODSIDECAR-130](https://folio-org.atlassian.net/browse/MODSIDECAR-130))
+
 ## Version `v2.0.13` (08.07.2025)
 ### Changes:
 * Refresh entitlement info upon failed entitlement check (MODSIDECAR-126)